+++ Classes/C2DXPaySDK/Android/com_mob_paysdk_cocos2dx_OnPayListener.h
@@ -8,21 +8,11 @@
 extern "C" {
 #endif
 
-/*
- * Class:     com_mob_paysdk_cocos2dx_OnPayListener
- * Method:    nativeOnWillPay
- * Signature: (ILjava/lang/C2DXString;Ljava/lang/Object;Lcom/mob/paysdk/MobPayAPI;)Z
- */
 JNIEXPORT jboolean JNICALL Java_com_mob_paysdk_cocos2dx_OnPayListener_nativeOnWillPay
-  (JNIEnv *, jobject, jint, jstring, jobject, jobject);
+  (JNIEnv *, jobject, jlong, jstring, jobject, jobject);
 
-/*
- * Class:     com_mob_paysdk_cocos2dx_OnPayListener
- * Method:    nativeOnPayEnd
- * Signature: (Lcom/mob/paysdk/PayResult;Ljava/lang/Object;Lcom/mob/paysdk/MobPayAPI;)V
- */
 JNIEXPORT void JNICALL Java_com_mob_paysdk_cocos2dx_OnPayListener_nativeOnPayEnd
-  (JNIEnv *, jobject, jint, jobject, jobject, jobject);
+  (JNIEnv *, jobject, jlong, jobject, jobject, jobject);
 
 #ifdef __cplusplus
 }
