+++ proj.android-studio/Cocos2dxBridge/src/com/mob/paysdk/cocos2dx/OnPayListener.java
@@ -15,16 +15,16 @@ public class OnPayListener<T> extends JavaCxxObject implements com.mob.paysdk.On
 
 	@Override
 	public boolean onWillPay(String s, T t, MobPayAPI mobPayAPI) {
-		int cxxThis = getCxxObject();
+		long cxxThis = getCxxObject();
 		return nativeOnWillPay(cxxThis, s, t, mobPayAPI);
 	}
 
 	@Override
 	public void onPayEnd(PayResult payResult, T t, MobPayAPI mobPayAPI) {
-		int cxxThis = getCxxObject();
+		long cxxThis = getCxxObject();
 		nativeOnPayEnd(cxxThis, payResult, t, mobPayAPI);
 	}
 
-	private native boolean nativeOnWillPay(int cxxThis, String s, T t, MobPayAPI mobPayAPI);
-	private native void nativeOnPayEnd(int cxxThis, PayResult payResult, T t, MobPayAPI mobPayAPI);
+	private native boolean nativeOnWillPay(long cxxThis, String s, T t, MobPayAPI mobPayAPI);
+	private native void nativeOnPayEnd(long cxxThis, PayResult payResult, T t, MobPayAPI mobPayAPI);
 }
